# API: allow disabling all callbacks