--- a/CHANGES.md
+++ b/CHANGES.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - CLI helper function now also supports normal (i.e. non-skorch) sklearn estimators
+- Disabling all callbacks is now supported (which allows reducing overhead,
+  which is especially relevant for small models).
 
 ### Changed