--- a/skorch/tests/callbacks/test_training.py
+++ b/skorch/tests/callbacks/test_training.py
@@ -1087,7 +1087,7 @@ class TestTrainEndCheckpoint:
 
 class TestNoCallbacks(TestCheckpoint):
     def test_no_callbacks(self, net_cls, data):
-        net = net_cls(callbacks=False)
+        net = net_cls(callbacks="disable")
         net.fit(*data)
-        assert net.callbacks is False
+        assert net.callbacks == "disable"
         assert net.callbacks_ == []