--- a/skorch/tests/callbacks/test_training.py
+++ b/skorch/tests/callbacks/test_training.py
@@ -1089,5 +1089,4 @@ class TestNoCallbacks(TestCheckpoint):
     def test_no_callbacks(self, net_cls, data):
         net = net_cls(callbacks="disable")
         net.fit(*data)
-        assert net.callbacks == "disable"
         assert net.callbacks_ == []