**What does this PR implement?**
It allows disabling all callbacks with `callbacks=False`.

I ran into a use case that uselessly spent 97% of the training time in various callbacks with the default `callbacks=None`. After this PR with `callbacks=False`, it only spends 5% of the total training time in various callbacks, a rather drastic speedup.