@BenjaminBossan 2020-11-16
For my better understanding, could you give more details about your use case. I can only imagine that the network and amount of data must have been really small for this to happen. Otherwise, it would indicate that one of our callbacks is much slower than it should be.

FYI, if I wanted to turn off all callbacks (only ever for debugging purposes, I would do this:

```python
net = NeuralNet(...).initialize()
net.callbacks_ = []
```