@stsievert 2020-11-17
Yes, my network is very small, and the data are between ~1KB to 1MB.

> FYI, if I wanted to turn off all callbacks (only ever for debugging purposes, I would do this:

Yes, I know. That's basically what this PR does. I'm submitting this PR because I probably spent an hour or two hours debugging the performance setback.