@BenjaminBossan 2020-11-18
I see the utility in this. What I'm unsure about is the use of `False` for this. It's really not intuitive why `None` would leave the default callbacks there but `False` removes them. I wonder if there is a better idea.

If that question is answered, I see two further points to improve:

- The documentation should be extended, it's not clear that this would touch the default callbacks.
- The test should be moved to `test_net.py`, because this is really a functionality of the net class itself.