@stsievert 2020-11-19
What's skorch's deprecation policy? I only implemented `callbacks=False` to not deprecate the API. I would prefer this API:

* `callbacks=None`: no callbacks whatsoever.
* `callbacks="auto"`: default callbacks.
* `callbacks=callback_list`: use those callbacks.