@BenjaminBossan 2020-11-19
For this change, I wouldn't break backwards compatibility. The utility would be quite low but everyone who uses the default callbacks would have to change their code, that's not worth it.

But maybe we can think of something else than `False` to completely remove all callbacks, including the defaults. Maybe `'no-defaults'`, `'empty'`, something like this?