@BenjaminBossan 2020-12-05
@stsievert what do you think of my suggestions?