@stsievert 2020-12-05
I agree – backwards compatibility should not be broken because unfortunately `callbacks=None` shows up in `NeuralNet(...).get_params()`.

I like `callbacks=False` because to me it says "callback existence is False". Out of the string options, I like `callbacks="ignore"` most because it answers the same question (but I like `callbacks=False` more). I don't like `callbacks in ["no-defaults", "empty"]` because that's asking for `callbacks=None`.