@BenjaminBossan 2020-12-12
> Out of the string options, I like `callbacks="ignore"` most

For me, `"ignore"` implies that there are callbacks, but they are ignored, whereas the implementation completely removes the callbacks. But in the end, I'm totally fine with this choice, what matters is that it's well documented.

Regarding your implementation, I would suggest a small change. Right now, you're checking at the end of `initialize_callbacks`, after all other mechanisms have already been performed unnecessarily. I would propose to perform this check at the beginning of the method and `return self` immediately afterwards. This is a bit more efficient and also more in line with the rest of the skorch code.