@stsievert 2020-12-15
I've implemented `callbacks="disable"` to allow disabling all callbacks. I've also better formatted the docstring for `callbacks`.