@BenjaminBossan 2020-12-19
Thanks so much @stsievert.

This is almost good to merge, just a tiny improvement: Your test now resides within `class TestNetSparseInput`, where it doesn't belong. I would suggest to put the test into the "section" about initialization: https://github.com/skorch-dev/skorch/blob/82a3da522a877c2bee8a276ddb4fdaebb8e44dbe/skorch/tests/test_net.py#L1422-L1424

That is really our fault for having such a huge test file, we should probably clean it up one day. It just grew organically to 3000 LOC and now it's hard to insert new tests there.