@BenjaminBossan 2020-12-19
> It is interesting how much overhead the callbacks have.

Could be worth investigating.