@BenjaminBossan 2020-12-19
I tried to run a few benchmarks but I had to try really hard to make disabling default callbacks matter. Here is what I came up with in the end:

```python
import numpy as np
from sklearn.datasets import make_classification
from torch import nn
from skorch import NeuralNetClassifier

# Only 10 samples, 5 features
X, y = make_classification(10, 5, random_state=0)
X = X.astype(np.float32)
y = y.astype(np.int64)

class MyModule(nn.Module):
    def __init__(self):
        super(MyModule, self).__init__()

        # very small net
        self.lin = nn.Linear(5, 2)
        self.softmax = nn.Softmax(dim=-1)

    def forward(self, X, **kwargs):
        X = self.softmax(self.lin(X))
        return X

net = NeuralNetClassifier(
    MyModule,
    max_epochs=500,
    lr=0.1,
    iterator_train__shuffle=True,
    train_split=False,
    #verbose=0,  # <= experiment 1
    #callbacks__train_loss=None,  # <= experiment 2 
)
net.fit(X, y)
```

As you can see, both data and the module are really really small, number of epochs are quite high. In these circumstances, disabling default callbacks reduces wall time from ~3.4 to ~1.7 sec.

What I found takes most time:

1. print log: setting `verbose=0` reduces time from 3.4 to 3.0 sec. This is not too surprising, given how it involves IO
2. train loss: disabling the calculation of train loss leads us from 3.0 to 2.2 sec, so pretty close to 1.7 sec without any default callbacks (note that I disabled the validation split)

My suspicion was that train loss calculation is slow because `history` access is slow. Indeed, almost all of the overhead is spent indexing into the batch logs: https://github.com/skorch-dev/skorch/blob/82a3da522a877c2bee8a276ddb4fdaebb8e44dbe/skorch/history.py#L204-L209

After a short time of tinkering, I couldn't find a way of accelerating these lines without making the code a lot more complicated. Thus I wonder if it's worth it to fix the issue or if we can consider this situation to be too marginal, especially since this PR will provide a solution that should satisfy most of these already rare situations.