@stsievert 2020-12-20
This more accurately mirrors my use case:

``` python
import numpy as np
from sklearn.datasets import make_classification
from torch import nn
from skorch import NeuralNetClassifier

class MyModule(nn.Module):
    def __init__(self, n_targets=85):
        super(MyModule, self).__init__()

        self.lin1 = nn.Linear(3, n_targets)
        self.lin2 = nn.Linear(n_targets, 2)
        self.softmax = nn.Softmax(dim=-1)

    def forward(self, X, **kwargs):
        X = self.softmax(self.lin1(X))
        X = self.softmax(self.lin2(X))
        return X


X, y = make_classification(n_samples=1000, n_features=3, n_redundant=1, random_state=0)
X = X.astype(np.float32)
y = y.astype(np.int64)

est = NeuralNetClassifier(
    module=MyModule,
    max_epochs=500,
    lr=0.1,
    iterator_train__shuffle=True,
    train_split=False)
```

Here are the timings on my machine:

* `net.fit(X, y)`: 12.4s
* `net.set_params(verbose=0)`: 10.0s
* `net.set_params(verbose=0, callbacks__train_loss=None).fit(X, y)`: 6.97s