@BenjaminBossan 2020-12-20
Thanks @stsievert, I can reproduce the relative numbers. With `verbose=0`, I get 20 sec, 11 of which are `history.__getitem__`. Unsurprisingly, it gets worse the longer the run is. I will investigate whether there are any worthwhile improvements there.

Meanwhile, could you please move the test as I indicated [above](https://github.com/skorch-dev/skorch/pull/719#issuecomment-748483376) so that we can merge?