@stsievert 2020-12-20
The test has been moved as suggested in https://github.com/skorch-dev/skorch/pull/719#issuecomment-748483376.